extract not found handler from server definition
